refactor: extract commands from shell
